requirements.txt
Django==4.2.16
Pillow>=10.0.0
requests>=2.31.0